[autotest] Add check_plugged_nodes_contain to check node plug status
In some tests, the test failed because of the malfunction of chameleon
plugger instead of Cras. To save our time for future debugging, we
should check the plug status and show them in the error message for
better triage and fix.

BUG=chromium:990233
TEST=Run on plugger function and not function host

Change-Id: I39960b41f469d0ffcca7eb509d1b4b3b4b2b584d
Reviewed-on: https://chromium-review.googlesource.com/1731236
Tested-by: En-Shuo Hsu <enshuo@google.com>
Commit-Ready: En-Shuo Hsu <enshuo@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>