[autotest] Select node after suspend to fix test flakiness
There is a known issue that will cause the selected device changed after
a suspend, which causing tests involving suspend flaky. We don't have
any plan or schedule to fix the issue recently. This is a temp fix before the
issue is address. This fix shouldn't hurt the purpose of the tests as
all of them are about testing the audio functions after a suspend
instead of the audio device selection.

BUG=chromium:990233
TEST=Run on lab dut

Change-Id: Icb3fec42c47ef1d49a1a2a9d79e045dbce522600
Reviewed-on: https://chromium-review.googlesource.com/1731237
Tested-by: En-Shuo Hsu <enshuo@google.com>
Commit-Ready: En-Shuo Hsu <enshuo@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>