network_WiFi_LinkMonitorFailure: don't fail upon lack of disconnect
This test is testing to ensure
(a) Shill's link monitor detects a link failure (e.g., failing ARP
    requests) within a certain timeout and
(b) the rest of the stack (driver, wpa_supplicant) responds
    appropriately when Shill requests a WiFi Reattach()

Reattach() is just a "Reassociation attempt":

  https://w1.fi/wpa_supplicant/devel/dbus.html#dbus_interface

which traditionally resulted in a disconnect/reconnect, but is now
widened such that some drivers support the reassociation attempt
directly, and therefore do not yield any disconnect. See
https://crbug.com/990012#c4 for some more information about this.

Anyway, the result is the 'iw event' only shows the "connected" event,
with no disconnect in the meantime. This was confirmed in a WiFi sniffer
packet capture. Since this behavior doesn't really violate (b), we
should loosen this restriction, and avoid failing the test just because
we didn't see a disconnect. A Reassociation should be as good as a
Deauth/Reassociation in terms of "un-sticking" the driver.

BUG=chromium:990012
TEST=run it on veyron_minnie-kernelnext

Change-Id: Iddc454cbfde00aa23e60de1cfdf9d956e79d2f9a
Reviewed-on: https://chromium-review.googlesource.com/1737468
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Neeraj Poojary <npoojary@chromium.org>