common-lib: ensure logging processes do not cause deadlock
The logic for logging sub-processes relies on closing the right fd and
the subprocesses getting that signal and exiting.
This breaks on the new servo_LabstationVerification test for a as of now
unidentified reason.
This change can be seen as proper robustness or a hack/workaround, but
essentially it wraps the normal logic in a timeout mechanism to ensure
that if the logging subprocess does not go down and appears to be stuck,
a signal is sent imploring it to exit.

BUG=chromium:970115
TEST=manual testing
test_that --autotest_dir . $DUT_IP servo_LabStationVerification.local

// observe the test not hang
// observe the logs show that the new mechanism had to be used

Change-Id: I7ff3c295483490200cb036f6c5a6f82116f3b7cc
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1735791
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Garry Wang <xianuowang@chromium.org>