cheets_CTS_P: Update waiver-candidate runner to test the new SkQP.
This is for upreving to the latest pie-cts-dev build, to test the
latest SKQP changes on Braswell P-experiment devices.

Since we already include full SkQP in the pattern (because the module
does not allow selecting a subset), only the CTS bundle is updated.

BUG=b:129918197
TEST=cheets_CTS_P.9.0_dev.arm.waivers

Change-Id: Ib7444b0b93e6f2608359ed71bf2dc2095726914c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1737408
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>