afdo_generate: Add a new board for async AFDO generation.
As part of the effort for moving AFDO generation out of PFQ, we need
to implement an async AFDO generation builder. To avoid interference
with current running AFDO generation on PFQ, we want to upload the
generate perf.data in the test to another GS bucket. We will use a
new board to minimize the changes to the script.

Exempt-From-Owner-Approval: The OWNERS file has a bug to identify the correct owners.
BUG=chromium:947345
TEST=None

Change-Id: I7bd3043f69456c937c2b222d8b727a8665f7a136
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1739747
Commit-Queue: Tiancong Wang <tcwang@google.com>
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>