autotest: Cleanup regex in crash_test
This makes the regexp more explicit so that it will only match lines we
want it to. We hadn't seen any further problems that I'm aware of from
the prior change, but there's an open bug to clean this up and it was
quick and easy to do.

BUG=chromium:923200
TEST=test logging_CrashSender

Change-Id: I6873bdc06830aabf337be522a1cd1b35183728c8
Reviewed-on: https://chromium-review.googlesource.com/1742605
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>