firmware_ECBootTime: add a delay before pressing the power button
As we changes the shutdown behavior for the mt8183 (ref cl:1660070),
the system will stay in S5 for 10s before entering G3. So, when we
do firmware_ECBootTime test, it will fail because the system is in S5
and the test script can't capture the string 'power state 1 = S5".
To adopt the EC change, this patch adds a delay to force the mt8183
enters G3 before pressing power button.

BUG=b:137605499
TEST=Pass the firmware_ECBootTime

Change-Id: I01f77a3c04a880c42fc3f39fed24a4677c063789
Reviewed-on: https://chromium-review.googlesource.com/1743610
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Commit-Ready: Zhuohao Lee <zhuohao@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>