atest host skylab_migrate command MVP
MVP functionality for `atest host skylab_migration` command.

`atest` post-this-CL is capable of migrating explicitly specified hosts and writes a JSON report of what it did to stdout, e.g.

```
{
    "duts": {
        "migrated": [
            "chromeos1-row4-rack9-host2"
        ],
        "needs_add_to_skylab": [],
        "needs_drone": [],
        "needs_rename": [],
        "not_locked": []
    },
    "failed_step": null,
    "locked_success": true,
    "plan": {
        "retain": [],
        "transfer": [
            "chromeos1-row4-rack9-host2"
        ]
    }
}
```
That being said, I had to remove a couple of bugs in order to get it to work end to end... and all of those fixes are in this diff.

That's why the diff is relatively large.

Non-MVP functionality such as specifying boards and pools and models has a draft implementation here, but hasn't really been tested.

Bug: chromium:989689
Change-Id: I8b573791d266851121ab3ed0aef94ad1474f213a
Reviewed-on: https://chromium-review.googlesource.com/1728425
Tested-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>