Remove dependency flag for video_AVAnalysis.
All the machines we need this test to run on have been moved to
a pool called: 'pool:cros_av_analysis' and the suite is spec'd to
run on that pool. As a result, we don't need to look for the label
'cros_av_analysis' as it is redundant. Removing so DUT selection
works properly.

BUG=None
TEST=Tested Locally

Change-Id: I1def02bfb488b8055a243b3989456a11218c501a
Reviewed-on: https://chromium-review.googlesource.com/1761652
Tested-by: Vinayak Suley <vsuley@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Max Timkovich <timkovich@chromium.org>