cr50_test: wait 2 * delay_reboot_to_ping for DUT
Wait 2 * delay_reboot_to_ping to wait for the DUT. It seems like we
aren't waiting long enough for the DUT to come up. Double the wait time,
so we do. The DUT will still reboot a couple of times if it doesn't
respond.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id3a17334300803a34fda88e839b78a547bcf80cd
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1764616
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>