autotest: PopAllUserProfiles() before calling configure_service_for_profile()
In light of recent Ethernet changes (and because ethernet_any is now
in many user profiles), running configure_service_for_profile() with
the profile argument set to the default profile path causes a check
to fail (the profile of the service to be configured must be equal to
or before the profile argument passed in) and an error ('Temporary
service configured but not usable') to be returned.

Call PopAllUserProfiles() so that the ethernet_any service's profile
will be the default profile.

BUG=None
TEST=test passes

Change-Id: Id05e07aa9deb926791770412bb5d3187089469f0
Reviewed-on: https://chromium-review.googlesource.com/1772688
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>