autotest: Move platform_HighResTimers to bvt-perbuild.
The test is ported to Tast already and running as a part of CQ.
Thus, it is redundant and consumes the resource to run it.
It is still a part of hwqual. Meanwhile, move this to bvt-perbuild
for minimum tracking of the test trajectory.

BUG=chromium:963325
TEST=None

Change-Id: I63e9899e7ff1486c0712902f5991473846a45f64
Reviewed-on: https://chromium-review.googlesource.com/1772915
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Laurent Chavey <chavey@google.com>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>