servo: log and skip any sort of uart capture failures
We have some DUTs deployed on v3 that have cr50 or usbpd overlays. These
duts will cause capture issues as they do have the control, but the
control is fake. This change fixes this issue by ignoring and logging
all uart capture issues in autotest.
In general, throwing an error here is not very helpful, as this is just
collecting more diagnostics.

This also fixes a second bug. If 'not_applicable' is the response to
turning on capture, then the capture failed, unlike the current code
indicates.

BUG=chromium:1000456, chromium:996328
TEST=None

Change-Id: Ie6289b32af9bb27116a71ba513c6dc6c42ca5e5a
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1783502
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Garry Wang <xianuowang@chromium.org>