cheets_CTS_{N,P}: Enables powercycling on P Bluetooth CTS.
Many of the boards that required this treatment have upgraded from N to P.
This cleanup step should be ported as well.

Besides, this CL modified the set up code to ignore any failures
on servo setup. This feature is, after all, only a nice-to-have
thing. Failing to have a servo is not a critical error. We shouldn't
give up running the test because of that.

BUG=b:126437757
TEST=cheets_CTS_P.9.0_r9.arm.CtsBluetoothTestCases

Change-Id: Ie4718fc1ad19f56d270ea03ce24d7acd01dc210b
Reviewed-on: https://chromium-review.googlesource.com/1809191
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>