bluetooth: Add debug info on first failure of test functions
Our test functionality is run with a _test_retry_and_log decorator
that re-tries a function in case of failure. After the second
failure, the test is marked as failed and the return value of
the function is printed. However, the return value of the first
run is equally, if not more important for debugging the cause
of failure, so this patch prints out the return value of the
first failure.

BUG=None
TEST=cause failures in autotest and observe
Change-Id: Id738d24f77366247b77d21b44ce71009b62df5ef
Reviewed-on: https://chromium-review.googlesource.com/1814223
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Yun-Hao Chung <howardchung@google.com>