firmware_TPMExtend: Fix test and add TPM 2.0 support
firmware_TPMExtend is supposed to test whether vboot correctly extends
the TPM PCRs on boot. Unfortunately, while the PCR test itself works
fine (for TPM 1.2), the test has a bug that causes it to "pass" even
when it found errors in the PCRs. This patch fixes this issue.

This test was written before the introduction of TPM 2.0, and since it
never failed we never realized that we needed to update it for that. Now
that it is again working correctly, it would fail on all TPM 2.0
devices, so this patch also adds the code to support those.

BUG=chromium:1005893
TEST=Ran on Kevin and Scarlet.

Change-Id: I829b87d0538b13f8753f103428b3b3ad6645ece2
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1817131
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>