pd_faft: Perform PDTester initiated items first, for safety reason
In firmware_PDVbusRequest test, it performs:
 * DUT initiated items, which PDTester is fixed to 20V and DUT initates
   the requests of 5V, 12V, or 20V (or the max input voltage DUT supports).
 * PDTester initiated items, which PDTester changes to support max 5V,
   12V, or 20V, and DUT reacts to request the max voltage.

The DUT initiated items is more risky. If the DUT max input voltage is
not 20V, like 12V, and the FAFT config is set wrong, it may negotiate to
to a voltage higher than it can support, that may damage the DUT.

The PDTester initiated items is safe. If the DUT max input voltage is
not 20V, it tries to negotiate the best. So it should not damage the DUT.
If it does damage DUT, this is also what we want to discover.

This CL swaps the order and performs the PDTester initiated items first.
If any item fails, possibly the FAFT config not set properly, it stops
the test and doesn't perform the risky DUT initiated items.

BUG=b:139372561
TEST=Ran firmware_PDVbusRequest. Made the PDTester initiated items
failed and checked it not perform the DUT initiated items.

Change-Id: I971c0383ca11754604a0ae57f6b074e7913809b2
Reviewed-on: https://chromium-review.googlesource.com/1823930
Tested-by: Wai-Hong Tam <waihong@google.com>
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>