Reduce the divergence between versions in controlfile generator
This CL is the preparation to unify the code of the scripts. This
CL makes the code of scripts same among the scriptsr, to make the
diff of this step visible in the git log.

I will extract the same code into one file after this is landed.

BUG=b:140981317
TEST=Run scripts

Change-Id: Ie00d9e4db1afba90300c8eba835e9155c88d1a2d
Reviewed-on: https://chromium-review.googlesource.com/1829598
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>