servo: introduce query for pd control on servo
This change adds the ability to query servo.py directly whether the
servo setup on the other end supports pd charging and control. This is
part of the effort to in the bug to shore up the autotest code to be
more capabilities oriented.

It also now removes that duplicate logic from the
server/cros/power/servo_v4_charge_utils.py, as well as some other small
maintenance on that code (reordering to get v4 role after checking v4 is
in use, deleting unused INA calls)

BUG=chromium:1009616
TEST=test_that --autotest_dir . $IP power_ServoChargeStress.3loops -b atlas

/.../results-1-power_ServoChargeStress.3loops   [  PASSED  ]

TEST=// remove charger cable
     test_that --autotest_dir . $IP power_ServoChargeStress.3loops -b atlas

... TestNAError: Servo setup does not support PD control. Check logs for
details.
// from logs
servo:1175| It appears that no charger is plugged into servo v4. Charger
port voltage: 5mv

Change-Id: Id2d7ae0723d110892c90ebbebd76c4f9b9dcf6ce
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1832587
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>