ssh_host: put more info in error descriptions
The error message is formatted "desctiption\n result_obj". The test
dashboards only show the first line of the failure, so we only see the
description of the ssh errors in the failure reasons. You have to look
in the test results to find the full failure reason. This change adds
more failure information to the description, so we can maybe see more
patterns from the test overview, and we wont need to go to the .INFO
or .DEBUG files to see the failure reason as much.

BUG=none
BRANCH=none
TEST=run unit tests

Change-Id: I9b251de3e117d95829d0cfdda882db199def4c16
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1833520
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>