faft-cr50: add wp g3 test
On some boards Cr50 can't deassert WP in G3, because the external WP
pullup isn't powered when the EC is off. Like the Rdd in G3 issue we
can't fix these hardware issues on released boards. We need to fix it on
new boards and track it on existing boards. This change adds a test to
verify Cr50 can disable WP in G3.

BUG=b:122538302
BRANCH=none
TEST=run on octopus and atlas

Change-Id: I1c2fe017c2cefe66a511d1d0c386447d4462f98b
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1835028
Reviewed-by: Dana Goyette <dgoyette@chromium.org>