[Autotest] Improving UI utils in Autotest
There is a corner case where a UI object will "show", but not be fully
ready for use. Rarely, when a test attempts to use the button it will
error out saying the button/object is not present. To help handle this a
built-in retry was added, which will check for the button to be set, and
retry up to 4 times if it is not.

TEST=policy_SecondarySigninAllowed
BUG=None

Change-Id: I8c1dc1d58ac8529e4d7d49405eed79d8a74200a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1841180
Tested-by: Derek Beckett <dbeckett@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ruben Zakarian <rzakarian@chromium.org>