autotest: cellular_Smoke: Ignore v6 DNS entries
Cellular currently has v6 support disabled. If the DNS server we
communicate with has AAAA entries for the hostname in question, a
corresponding socket.getaddrinfo call would have some entries
containing v6 server addresses. This change allows for cellular_Smoke
to ignore v6 entries when calling CheckInterfaceForDestination.

Note that doing this opens the possibility for FetchUrl to use a
different physical interface to communicate with the server over
v6. We can get away with this because cellular is the lowest priority
physical interface, meaning that CheckInterfaceForDestination would
fail if WiFi or ethernet could potentially be used. In the long-term,
something like CL:1404384 would allow for the enforcement of either v4
or v6 for communication with the server and for
CheckInterfaceForDestination.

BUG=b:141320188, chromium:932661
TEST=`test_that $IP cellular_Smoke.fibocom-l850gl` passes, ensuring
     that we are *not* auto-connecting to a different network (e.g. VPN
     or WiFi) during the test.

Change-Id: I2fec4baeecb405eb038b8d20c25bad9c11e67794
Reviewed-on: https://chromium-review.googlesource.com/1845687
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>