cheets_[CG]TS: Factor out the qual suite name.
This is to reflect the control file change in CTS_Q collect-test,
so that it is generated under suite:arc-cts-q, not in the fixed
arc-cts nor arc-cts-qual.

Along the way, the CL fixes the suites for P collect-tests.

BUG=b:142095998
TEST=regenerated control files

Change-Id: I578bc9d5f7aeea8b931ee617a934ee17744ddc66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1863269
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>