[Autotest] Reducing flakiness in policy_ScreenBrightnessPercent
There is some flakiness when running this test with the 50% case. It
appeared to be caused by the backlight not settling in the
time.sleep(1).
Changed this to a while loop for 5 seconds until the previous 2 values are
very close then break.
Additionally, some logging was added to help future debugging.
BUG=None
TEST= 25 iterations of this, no failures.

Change-Id: Ic19d95750688dbcbf5d3704b729106038c07524c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1864386
Reviewed-by: Ruben Zakarian <rzakarian@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>