autotest: greatly reduce skylab_test_push behavior
Prior to this, skylab_test_push was running through a variety of
now-obsolete flows.

Replace it's behavior with an attempt at a repair task of a single DUT
in the staging lab. This is adequate for ensuring basic sanity of
autoserv and the autotest repo, and provides some basic coverage that
lab platform is working.

Follow-up: make skylab_test_push more robust by not hardcoding the DUT
to repair.

BUG=chromium:1014685
TEST=local run of bin/skylab_test_push

Change-Id: If6d558eafc771f1119f53bfd939412a5784914eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1867175
Reviewed-by: Allen Li <ayatane@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>