faft: Separate the run_shell_command method that returns a result.
In CL:1863838, the behavior for block=True changed, breaking tests.
Before, it waited for process to exit; after, it waited for stdout
to be read as well.

This commit changes the method back to its old behavior, and adds a
separate method to give a result with stdout, stderr, and returncode.

BUG=b:147227865
TEST=Run firmware_FAFTRPC.system

Change-Id: Ie516a8baed2b5d1a19ca764e62f037e224da16e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1988732
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>