autotest: network_WiFi_VisibleScan: add VHT80 option
Adding a test version to network_WiFi_VisibleScan which forces the
AP's channel to an 80 MHz wide channel. This is part of the effort
to enhance the VHT testing capabilities available in autotest.

BUG=chromium:827829
TEST=`test_that -b caroline ${DUT_IP}
network_WiFi_VisibleScan.width_5VHT80`.
Run `watch -n0.1 "iw wlan0 station dump | grep 'tx bitrate'"`
on the device and check that the bitrate is "80MHz VHT-NSS" for
the 5VHT80 version and not that for the 11b version of the test.

Cq-Depend: chromium:1877829
Change-Id: I9b0e0a86947d22ef66bdaa447b23ff154ac88727
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1879476
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Commit-Queue: Kirtika Ruchandani <kirtika@chromium.org>