cheets_CTS/GTS: simplify the command generation logics
Previously the command generation logics are devided into
for-cts-instant and for-non-instant. This is because
the CTS Instant doesn't support --include-filter option.

This CL changes the rule to unify the logics:
- For the test with one module: use --moduel option
- For the test with multiple modules: use --include-filter option

This rule is ok since the CTS instant only has the tests
with one module.

And this CL also changes the controlfiles accordingly.

BUG=b:142095998
TEST=Ran the CTS/GTS generators

Change-Id: I63160cb7e491bae97fc62a8f971e1321f6af3eac
Reviewed-on: https://chromium-review.googlesource.com/1880613
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>