cheets_CTS/GTS: add a feature to cache the downloaded bundle file
This CL adds a '--cache_dir <DIRNAME>' flag for the generator
script. This flag caches the downloaded bundle file, or uses
the cached bundle file if available. So we can reduce the network
usage and the time at the second try and after.

This is a kind of the retry of crrev.com/c/1877798, which broke
controlfiles since they need the actual URI in the controlfile.

BUG=b:142095998
TEST=Ran the CTS generators with and without the flag

Change-Id: I354e0bf96ed49c70201575d5b37459aa8bb10ae2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1880390
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>