pd_faft: Don't create PDTester host if (hostname:port) are the same as Servo host
If the hostname and the port of PDTester are the same as Servo, will use
the same Servo object as PDTester. We don't need to create another
PDTester object, that causes a problem in RpcServerTracker. Because if
calling RpcServerTracker to create a same (hostname:port) ssh tunnel, the
previous-created ssh tunnel will be disconnected and then create a new
one; as a result, the reference to the previous ssh tunnel will get
lost.

Also change to use the default hostname and port from Servo host,
instead of the Servo arguments.

BUG=chromium:1017343
TEST=Ran a PD FAFT test with the same args of PDTester and Servo.

Change-Id: If4394a40a4a57f6552ad16955b38afef0f81450f
Reviewed-on: https://chromium-review.googlesource.com/1881727
Tested-by: Wai-Hong Tam <waihong@google.com>
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Matthew Blecker <matthewb@chromium.org>