faft: In consecutive reboot test, shut down in background.
Previously, the test triggered a race between the system shutdown
and the reply to RunShellCommand("shutdown -P now").

This could result in a 20 minute hang, followed by an attempt to
ssh in to the DUT that was just shut down.

Now the test executes the shutdown in the background, and catches
and logs any ssh errors caused by the shutdown.

BUG=b:141519692
TEST=Run firmware_ConsecutiveBoot.dev.100

Change-Id: Ice769cd7bd7a21e74ad5c25db8e8ca2cbcc42492
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1881777
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>