power_BatteryDrain: Add client utility test
This test drains the battery quickly, as a utility for other tests. It
turns the screen and keyboard backlight up as high as possible, and then
opens Chrome to a WebGL heavy webpage. It requires that AC power is not
supplied, so you will have to turn it off via servo in a server test.
You can specify a battery threshold to drain to, and a timeout in
seconds.

On my Sarien device this puts the battery drain rate at about 35 watts,
as compared to about 6 when sitting idle at the login screen with a
medium backlight. In conjunction with a USB power load tester
(https://amzn.com/B01LKT0OTQ) that draws an additional 10 watts, the DUT
drains at about 1% a minute. For the policy_DevicePowerPeakShift and
policy_DeviceAdvancedBatteryChargeMode tests that I am writing, which
require the DUT to be at about 95% battery (so they can charge), this
only takes about 5 minutes.

BUG=b:139201701, b:138940522, b:143173004
TEST=test_that -b sarien \
--args='drain_to_percent=90 drain_timeout=300' $DUT power_BatteryDrain

Change-Id: Id7373d7b929cfdf3d71ffde37ef3d5dea501bfc7
Reviewed-on: https://chromium-review.googlesource.com/1867569
Tested-by: Nick Crews (Nov 1 last day) <ncrews@chromium.org>
Commit-Ready: Daniel Campello <campello@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Shawn N <shawnn@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>