Remove self-referential content-refresh line
In the beginning, go/faft was a Google Doc.
Then we put its contents into a markdown file, and added a line
declaring when the content had last been refreshed from go/faft.
Then go/faft was updated to point to the markdown file, but it still
contained that line that talked about go/faft.

This CL removes that line.

TEST=None
BUG=None

Change-Id: I2f597724adb41e531304961618341194bf0981cd
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1888393
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>