autotest: move hardware_StorageFio functionality to a base test
We want to add harware_StorageFioOther as a derivative of hardware_StorageFio,
so this moves the functionality to a base test class called FioTest, which can
then be used in both client tests.

BUG=b:143568768
TEST=Ran hardware_StorageFio, hardware_StorageFioOther.test,
hardware_StorageSuspendStress.test, and hardware_StorageSuspendStress.test_other
on different systems.

Change-Id: I6cc284904b3306c4abc1f845d0adc09e934edd50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1888901
Commit-Queue: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>