faft: check power state in the test, not in servod
Previously, this code ran and waited for 'powerinfo' with a regexp to
match the desired power state, so the state being wrong would look like
a failure to respond to the command.  Now the method returns the state
name, and the test itself checks that it matches the expected value.

The old logic didn't explicitly define a timeout, but it turned out
to be 3 seconds per try.

BUG=b:141519692
TEST=Run a variant of firmware_ConsecutiveBoot
Change-Id: Ide24f803bef4f85e792803813418af6de23e5dbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1895163
Tested-by: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>