autotest: change hardware_StorageFio to allow for other devices
Some other (non-root) devices that we want to test will not be partitioned.
Change the regular expression to accomodate the potential lack of partition
numbers and don't set the filename to be the first partition of the external
drive if it doesn't exist.

BUG=b:143568768
TEST=Ran hardware_StorageFio, hardware_StorageFioOther.test,
hardware_StorageSuspendStress.test, and hardware_StorageSuspendStress.test_other
on different systems.

Cq-Depend: chromium:1888901
Change-Id: I15316d9d52e392271a3d10d25b73ab90063847e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1894120
Commit-Queue: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>