power_status: Improve battery_charging()
If either linepower or the battery thinks were charging, then we
probably are. Some devices don't give the status field of linepower, but
they do give the status field of the battery. My Sarien for instance
does not have linepower.status, and this was the impetus for this CL,
because the old ac_charging() function did not work on Sarien (it said
not charging, when in fact it was charging).

If this breaks any tests by changing the behavior, then those tests are
flawed: They thought that the DUT was not charging, when it really was.

While were at it, rename ac_charging() to battery_charging(), because
that's what really is happening, and simplify the |= logic.

BUG=b:139201701, b:138940522, b:143173004
TEST=Only one test seems to transitively use this function:
test_that -b sarien $DUT power_ProbeDriver.probe_ac
and this is still succeeds.

Change-Id: I49de75afd5b3fe0891623da5219fb38009c6232d
Reviewed-on: https://chromium-review.googlesource.com/1888788
Tested-by: Nick Crews (Nov 1 last day) <ncrews@chromium.org>
Commit-Ready: Daniel Campello <campello@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>