bluetooth: add test for querying device connection info
The D-Bus API GetConnInfo() on interface org.chromium.BluetoothDevice
was not included in the autotest. This absence caused a bug to slip
through.

When a device is connected, calling this method should return the
connection info, otherwise it would return an error. For more info
on the API please refer to third_party/bluez/doc/chromium-api.txt.

BUG=chromium:1019443
TEST=Run test bluetooth_AdapterQuickSanity locally

Change-Id: I69b7939140225e7e9c1138c0d9c7933d1b6ccba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1900951
Tested-by: Archie Pusaka <apusaka@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>