autotest: remove blktrace profiler and btreplay test
blktrace is now included in test images, so there is no need for
this profiler.

btreplay is not being used and is broken in it's current form. If
we want to replay traces, we can add that functionality to current
tests, but there is no use for this test as is.

BUG=chromium:903477
TEST=None

Change-Id: Iaa8ba0a5089272af2ef057817aae8e6f2e915446
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1902061
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>