bluetooth: improve stability in connection
In test_device_is_connected, it runs a for loop to check if property 'Connected'
in 'org.bluez.Device1' is True. But the property only reflects the
status of physical link, it may have chance that the phycical link is
connected for a small period of time but the connection failed after a
while.

The patch fix it by sending a mouse/keyboard event after checking the
property 'Connected'. If the connection does not success, the HID report
can not be received.

The Classic batch is also included in the quick sanity package since the
batch is stable after this patch.

BUG=b:143673942
TEST=run quick sanity stress locally

Cq-Depend: chromium:1902337
Change-Id: I84dace0f269fbaab3a7d1e1ce567be25b6f7230c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1905308
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Tested-by: Yun-Hao Chung <howardchung@google.com>
Commit-Queue: Yun-Hao Chung <howardchung@google.com>
Auto-Submit: Yun-Hao Chung <howardchung@google.com>