power_suspendStress: Use suspend_for_sec instead of wakeup_timeout.
SuspendStress test intends to test the stability of suspend/resume on a
given device. Currently we set a alarm in the future and ask powerd to
suspend the device. powerd then send d-bus messages to interested
daemons about the imminent suspend and suspends the device when those
daemons indicate that they are ready. If this takes more time than
expected, the above alarm can go off either before we suspend or
when kernel is in the process of suspending resulting in spurious
suspend failures.

Instead this change asks powerd to set the alarm just before suspending.
This reduces the spurious failures and increases the reliability of the
test.

However, this means we do not exactly know when in the future the alarm is
set to. Thus we cannot reliably measure the board up time as part of
this test anymore. For measuring board_up time, power_Resume test should
be used.

BUG=chromium:1017271, b:141497269
TEST=run power_resume before and after and make sure total_up time comes
down.

Change-Id: Ia000b57fd124ab08387d8845ced17e3eab3a1365
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1904456
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Commit-Queue: Ravi Chandra Sadineni <ravisadineni@chromium.org>