[autotest] Make some audio tests only dep on audio_box
We have an agreement that the audio_box should always be with
audio_board and audiojack. It is redundant to have these labels along
with audio_box.

Also, as part of audio test improvement plan, this CL also moves other
tests depend on audio_box to unstable to limit their run. The intention is to
have only essential audio tests run in audio box before we make them stable
enough to catch regressions. They are pretty flaky and hard to maintain
recently because of our current setup.

BUG=b:140781867
TEST=None

Change-Id: I8b5a06c1aa5db002c5b6801dbff22b5e14913675
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1911521
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>